pipelined set/get

+++ src/main/java/org/skife/memcake/Opcodes.java
@@ -3,5 +3,5 @@ package org.skife.memcake;
 class Opcodes {
     static final byte set = 0x01;
     static final byte add = 0x02;
-    static final int get = 0x00;
+    static final byte get = 0x00;
 }
